+++ database/migrations/2026_01_17_132320_create_payments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->foreignId('order_id')->constrained();
-            $table->unsignedBigInteger('amount');
+            $table->float('amount');
             $table->timestamps();
         });
     }
